+++ examples/trace.cxx
@@ -12,10 +12,10 @@
 int trace(int const     n,
           CTF_World    &dw){
   int rank, i, num_pes;
-  long long np;
+  int64_t np;
   double * pairs;
   double tr1, tr2, tr3, tr4;
-  long long * indices;
+  int64_t * indices;
   
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   MPI_Comm_size(MPI_COMM_WORLD, &num_pes);
